requirements.txt
absl-py==1.4.0
adversarial-robustness-toolbox==1.13.1
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
art==5.8
astor==0.8.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.0.1
cleverhans==3.0.0
cloudpickle==2.2.1
cycler==0.11.0
datasets==2.10.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
dm-tree==0.1.8
easydict==1.10
entrypoints==0.4
fastjsonschema==2.16.3
filelock==3.9.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
gast==0.5.3
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.8.0
huggingface-hub==0.12.1
idna==2.10
imageio==2.27.0
importlib-metadata==6.0.0
importlib-resources==5.12.0
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter==1.0.0
jupyter-client==7.4.9
jupyter-console==6.6.3
jupyter-core==4.12.0
jupyter-server==1.23.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
keras==2.11.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==2.0.5
mnist==0.2.2
multidict==6.0.4
multiprocess==0.70.14
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==2.6.3
nose==1.3.7
notebook==6.5.3
notebook-shim==0.2.2
numpy==1.21.6
packaging==23.0
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil-resolve-name==1.3.10
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.22.0
psutil==5.9.4
ptyprocess==0.7.0
pyarrow==11.0.0
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.14.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==25.0.1
qtconsole==5.4.1
QtPy==2.3.0
requests==2.25.1
responses==0.18.0
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-probability==0.19.0
termcolor==2.2.0
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2021.11.2
tinycss2==1.2.1
torch==1.13.1
torchattacks==3.3.0
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
typing-extensions==4.5.0
urllib3==1.26.14
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.5
wrapt==1.15.0
xxhash==3.2.0
yarl==1.8.2
zipp==3.15.0
